clear_and_send_keysに書き換えてみた

+++ rikson_simulation.py
@@ -79,21 +79,22 @@ jj = int(float(sys.argv[1]))
 
 for symbol in ['USDJPY', 'EURUSD']:
     for i in range(20):
-        rik = 0.00005*(i+1)
-        for j in range(2):
-            son = 0.00005*((j+1)*jj)
-
-            simulator = Simulator(symbol, rik, son)
-            kane, asset = simulator.run_simulation()
-
-            print(symbol, rik, son, kane)
+        rik = 0.00001*(i+1)
+        for j in range(4):
+            son = 0.00005*(j*10+jj)
 
             base_dir = 'datas/simulation/{}'.format(symbol)
             os.makedirs(base_dir, exist_ok=True)
             save_dir = base_dir + \
                 '/m{}_rik{}_son{}.npy'.format(1,
-                                              "{:.4f}".format(rik),
-                                              "{:.4f}".format(son))
+                                              "{:.5f}".format(rik),
+                                              "{:.5f}".format(son))
 
-            np.save(save_dir, np.array(asset, dtype='float32'))
+            if not os.path.exists(save_dir):
+                simulator = Simulator(symbol, rik, son)
+                kane, asset = simulator.run_simulation()
+
+                print(symbol, rik, son, kane)
+
+                np.save(save_dir, np.array(asset, dtype='float32'))
 # %%
